+++ simulation/simple_sensor_simulator/include/simple_sensor_simulator/sensor_simulation/sensor_simulation.hpp
@@ -33,18 +33,11 @@ public:
     const double current_simulation_time,
     const simulation_api_schema::LidarConfiguration & configuration, rclcpp::Node & node) -> void
   {
-#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
     if (configuration.architecture_type() == "awf/universe") {
       lidar_sensors_.push_back(std::make_unique<LidarSensor<sensor_msgs::msg::PointCloud2>>(
         current_simulation_time, configuration,
         node.create_publisher<sensor_msgs::msg::PointCloud2>(
           "/perception/obstacle_segmentation/pointcloud", 1)));
-    } else
-#endif
-      if (configuration.architecture_type() == "awf/auto") {
-      lidar_sensors_.push_back(std::make_unique<LidarSensor<sensor_msgs::msg::PointCloud2>>(
-        current_simulation_time, configuration,
-        node.create_publisher<sensor_msgs::msg::PointCloud2>("/perception/points_nonground", 1)));
     } else {
       std::stringstream ss;
       ss << "Unexpected architecture_type " << std::quoted(configuration.architecture_type())
@@ -58,19 +51,11 @@ public:
     const simulation_api_schema::DetectionSensorConfiguration & configuration, rclcpp::Node & node)
     -> void
   {
-#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
     if (configuration.architecture_type() == "awf/universe") {
       using Message = autoware_auto_perception_msgs::msg::PredictedObjects;
       detection_sensors_.push_back(std::make_unique<DetectionSensor<Message>>(
         current_simulation_time, configuration,
         node.create_publisher<Message>("/perception/object_recognition/objects", 1)));
-    } else
-#endif
-      if (configuration.architecture_type() == "awf/auto") {
-      /* Autoware.Auto does not currently support object prediction however it is
-         work-in-progress for Cargo ODD msgs are already implemented and
-         autoware_auto_msgs::msg::PredictedObjects will probably be used here
-         topic name is yet unknown. */
     } else {
       std::stringstream ss;
       ss << "Unexpected architecture_type " << std::quoted(configuration.architecture_type())