+++ simulation/traffic_simulator/include/traffic_simulator/traffic_lights/traffic_light.hpp
@@ -15,10 +15,7 @@
 #ifndef TRAFFIC_SIMULATOR__TRAFFIC_LIGHTS__TRAFFIC_LIGHT_HPP_
 #define TRAFFIC_SIMULATOR__TRAFFIC_LIGHTS__TRAFFIC_LIGHT_HPP_
 
-#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
 #include <autoware_auto_perception_msgs/msg/traffic_signal.hpp>
-#endif
-
 #include <iostream>
 #include <limits>
 #include <stdexcept>
@@ -78,7 +75,6 @@ public:
   auto colorChanged() const { return color_changed_; }
   auto arrowChanged() const { return arrow_changed_; }
 
-#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
   explicit operator autoware_auto_perception_msgs::msg::TrafficSignal() const
   {
     autoware_auto_perception_msgs::msg::TrafficSignal traffic_light_state;
@@ -102,7 +98,6 @@ public:
 
     return traffic_light_state;
   }
-#endif
 
 private:
   std::unordered_map<TrafficLightColor, geometry_msgs::msg::Point> color_positions_;